Use given default as default per type in activity vertex attributes
The functions add.vertex.attribute.first.activity and
add.vertex.attribute.active.ranges both accept a parameter default.value. This
given value is interpreted as default value per type and author (not as default
attribute for vertices for which no data is available). It is distinguished from
the constructed vertex default and added to the vertex attributes in the
corresponding helper functions.

Signed-off-by: Klara Schlueter <schluete@fim.uni-passau.de>