Fix extraction functions regarding handling of empty networks
If a network is empty, the function
'extract.bipartite.network.from.network' results in an error, as the
sub-setting of the edges does not work. This can happen in two cases:
1) the network does not contain vertices and 2) the network does not
contain the edge attribute 'type'.

The problem is fixed by adding according checks to the function
implementation. Now, an error is thrown for Case 2), while Case 1) is
handled by doing essentially nothing.

Tests for all extraction functions 'extract.*.network.from.network' are
added to properly identify potential regressions in the future.

This fixes #155. Props to @bockthom for reporting this.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>