Fix tests for 'add.vertex.attribute.author.role.simple'
After applying commit 9294a37d98f9ff3d14756d56300b0d171f3f3b4c, the
tests for the function 'add.vertex.attribute.author.role.simple'
misbehave as the thresholds for the classification shifted.

To fix the problem, the expected author roles and, thus, vertex
attributes are adapted to the real values after altering the threshold
in the said commit. The expected output was verified by debugging the
actual data and classification.

This relates to PR #156.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>