Always add mandatory vertex and edge attributes
When constructing any type of network by calling the corresponding
methods on a 'NetworkBuilder' instance, the mandatory vertex and edge
attributes need to be added in any case – and also all edge attributes
configured in 'Networkconf$get.value("edge.attributes")'. Before, this
did not work for all types of networks.

The crucial functionality to enable the addition of edge and vertex
attributes to existing (empty) networks is added by the function
'add.attributes.to.network', which is able to perform this step.
The function 'create.empty.network' is refactored to use this new
function properly when 'add.attributes = TRUE'.

The fix for the edge attributes consists of the following steps:
1) For consistent access to available edge attributes, the method
'ProjectData$get.data.columns.for.data.source' is added to return a list
of data classes, accordingly with the available data columns as names.
2) The function 'construct.network.from.edge.list' gets a new parameter
'available.edge.attributes' to which all affected network-construction
methods pass the appropriate data from Step 1). Internally, the function
'add.attributes.to.network' is called when the edge list is empty (this
is not needed otherwise).
3) The function 'add.edges.for.bipartite.relation' is extended
analogously to the function in Step 2) to get the available attributes.

The fix for the vertex attributes consists of the following steps:
A) In the function 'construct.network.from.edge.list', the vertex
attribute 'name' is added when the list of vertices is empty. All other
attributes are added in the respective methods
'NetworkBuilder$get.*.network'.
B) As bipartite networks are constructed differently, a corner-case for
an empty list of vertices is added to the method
'NetworkBuilder$get.bipartite.network' to add the mandatory vertex
attributes.

Tests are added for network construction with and without data being
available for the vertex and edge attributes (see file
'tests/test-networks.R').

There some minor adjustment to the code on the way:
- The list of vertices returned from the function
'construct.edge.list.from.key.value.list' is now unique.
- The first two columns of the edge list constructed in the function
'construct.edge.list.from.key.value.list' are now named "from" and "to".
- The construction of the additional edge attributes in the function '
add.edges.for.bipartite.relation' is now robust against an empty edge
list.
- Some tests in the files 'tests/test-networks.R',
'tests/test-networks-author.R', and
'tests/test-networks-multi-relation.R' are fixed regarding the
construction of the edge attributes, caused by to implicit changes.

This fixes #154.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>