Fix wrong behaviour of 'get.author.class' when using 'result.limit'
The use of 'result.limit' should not alter the classification results however
this was the case in the old version. This is now changed to work as intended.

A special if-statement within the function checks if the minimum of the
variable 'buffer.value' does not exist (i.e., is infinite). If it is infinite,
another local index variable was set to 0 which resulted in all developers but
the first one being classified as peripheral. However, what should happen is
that everyone is classified as core. Therefore, this index variable must be set
to the number of authors instead.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>