Add source filter to issue read method
The method 'read.issues' can now filter if jira, github or both issues
are read in. Therefore the project configuration gets the new parameter
'issues.from.source'. Default setting is that both, jira and github
issues, are read in. The issue data from all sources is still parsed to
the same data-frame, but additionally the source information is now part
of the issue.id and saved in an extra column in the data-frame.

Signed-off-by: Anselm Fehnker <fehnker@fim.uni-passau.de>