Adjust tests to issue source filter
The tests containing issue data are updated to the possibility to filter
if only jira, only github or both issue data shall be read in. Therefore
the issue data gets split in two files, 'issues-jira.list' for the jira
issues and 'issues-github.list' for the github issues. Also two new
issues read tests are added, one that reads only the jira issues and one
that reads only the github issues. In all already existing issue tests
the 'issue.id' is updated.

Signed-off-by: Anselm Fehnker <fehnker@fim.uni-passau.de>