Do not redundantly initialize data sources when splitting
The first essential part of the splitting functionality is the
initialization of all needed data sources, both the main data sources
and the additional ones. While the main data sources are split, the
additional data sources are left unaltered.

The getters for the additional data sources are called right upon the
(parallel!) creation of 'RangeData' objects, leading to potentially
redundant data retrievals from disk. To avoid this problem, the
additional data sources are retrieved and initialized right at the start
of the splitting process, without calling the getter multiple times.

In addition, the data initialization of the 'RangeData' object is
slightly refactored to be prepared for further data sources.

Thanks to @bockthom and @ecklbarb for pointing out the problem. Thanks
to @bockthom for discussing solutions.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>