Fix reading of synchronicity data
When reading synchronicity data, the data is not properly added to the
commit data – even if the 'ProjectConf' value 'synchronicity' is
enabled. To fix this, the methods 'ProjectData$get.synchronicity' and
'ProjectData$set.synchronicity' are streamlined in resemblance to the
mail/commit-related methods.

Moreover, the data column that is added to the commit data is renamed to
be 'synchronicity' (formerly, 'synchronous') to establish analogism to
the PaStA data. Additionally, the edge attribute has been named
'synchronicity' before (rendering it praktically broken without this
very patch!). The README and the test suite are adapted accordingly.

Furthermore, the constants 'SYNCHRONICITY.LIST.COLUMNS' and
'SYNCHRONICITY.LIST.DATA.TYPES' as well as the function
'create.empty.synchronicity.list' are added, analogously to the
functions and constants for the other data sources.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>