Read PaStA and synchronicity data only if enabled
To avoid unnecessary reading from disk for the PaStA and synchronicity
data, the data is now only read when and if the 'ProjectConf' variables
'pasta' and 'synchronicity', respectively, are enabled.

Furthermore, the check for a null value is added to the method
'ProjectData$set.pasta' as it is present for all other data sources.

Thanks to @ecklbarb and @bockthom for pointing this out.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>