Fix bulk module
The functions in the bulk module did not work any more as they did not
use the `NetworkBuilder` class to construct networks, but the
`RangeData` class. As the network construction functions and classes
have already been changed some time (years) ago, this needs to be fixed
for the bulk-module functions to work again.

In this commit, the bulk module gets updated in such a way that all the
function in the module are usable again.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>