Replace 'get.commits' call within function 'get.commit.data'
Replace 'get.commits' call with a 'get.commits.filtered' call in the function
'get.commit.data' and adjust affected test cases accordingly.

This change was made because at each place where 'get.commit.data' is used, it
is not wanted that successive calculations work with unfiltered commits.

This relates to issue #70.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>