Format and simplify code
The review from @clhunsen is applied. This includes code formatting and
code simplifying. So the command to combine the issue data from
different sources and the way how the "closing.date" is parsed are
changed. Through these changes the if statement that checks if the issue
date is not empty can be removed.

Signed-off-by: Anselm Fehnker <fehnker@fim.uni-passau.de>