Fix crash of function 'get.author.class' when classifying zero authors
Fix crash behaviour of function 'get.author.class' which occurred whenever a
zero-row dataframe was passed in the parameter 'author.data.frame'.

The fix is realized without a parameter check but instead with two slight
manipulations to the existing code.
 - The expression '1:author.class.threshold.idx' is replaced with
   'seq_len(author.class.threshold.idx)' to always produce a integer vector of
   length 'author.class.threshold.idx', specifically in the case of
   'author.class.threshold.idx' being zero which occurrs when a zero-row
   dataframe is passed through the parameter 'author.data.frame'.
 - The function 'sapply' shows a strange behaviour whenever an empty vector is
   passed as the first argument (in this case 'author.cumsum'). It always
   returns vectors having the same length as the first argument, however, when
   the first argument is a vector of length zero, it returns an empty list
   instead of an empty vector. Therefore, an 'as.logical' statement is
   added to ensure that there is always a (logical) vector being returned.

The two above mentioned changes allow the function to handle zero-row dataframes
being passed correctly without additional parameter checks.

In addition, a call to 'suppressWarnings' is used to hide the warning that was
output when the function 'min' gets called on an empty vector. The warning
informed about 'min' returning an infinity value since no minimum value could be
found in the empty vector, however, this special case is handled in the
following 'if' statement anyway, so there is no need to show this warning to the
user.

This fixes #164.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>