Add vertex attribute also to empty networks
Given the corner case that an empty network is given to the function
'add.vertex.attribute.*' or none of the vertices in the network is
assigned a value, the vertex attribute is now added manually and by
force using 'add.attributes.to.network'. The default value is then
assigned as defined by the immediate call to 'add.vertex.attribute.*'.

Two test cases are added:
- addition of vertex attributes to empty networks, and
- addition of vertex attributes to non-empty networks, but with usage of
  the default value (adaptation of first new test case).

Additionally, to foster readability inside the function
'add.vertex.attribute', the local variable 'attr.df' is renamed to
'attrs.by.vertex.name' – since it is no data.frame.

This fixes #165.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>