Add additional check for empty data to 'get.author.class'
For the case that a data.frame with less than two columns is given to
the function 'get.author.class', the input data is reset to coincide
with the specification given in the function documentation.

This is related to #164 and 8060caa4930ef05a48b59a328ebf928a64109294.

Additionally, a short improvement to the documentation of
'get.author.class', as the column denoted by 'calc.base.name' does not
necessarily need to be the second column.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>