Add possibility of scaling editor type plot to 100%
A parameter plot.percentage is added to the function
plot.commit.editor.types.by.author. If true, the barplot shows the relative
numnber of differently edited commits per author: each bar in the barplot
(representing the commits of one editor) is scaled to 100%. Otherwise, the
absolute number of commits per author is shown in the plot.

Signed-off-by: Klara Schlueter <schluete@fim.uni-passau.de>