Order bars in plot of plot.commit.editor.types.by.author
The bars in the plot are ordered as follows (criteria are listed with
descending priority): the editor who committed the most commits that where
authored by someone else, is represented by the rightmost bar. The editor who
committed the most commits that where also authored by him is represented by the
rightmost bar. The remaining editors (who are only authors) are sorted by the
number of commits.

Signed-off-by: Klara Schlueter <schluete@fim.uni-passau.de>