Add preliminary methods to calculate EDCPTD centrality
Methods that enable to calculate the EDCPTD centrality for a project are
added to a provisional file 'util-tensor.R'. This include methods that
retrive the needed single-layer developer networks, a method that builds
a forth-order tensor using this single-layer networks and e method that
calculates the EDCPTD centrality using this tensor.

Signed-off-by: Anselm Fehnker <anselm@muenster.de>