Ensure correct ordering and handle PaStA duplicates
Previously, it was possible that in certain use cases the ordering of mails and
commits could end up mixed up because a sorting operation was lacking. This is
added in this commit.

For PaStA, it is possible that after filtering patchstack mails, duplicated
PaStA rows can occur which are filtered from now on.

Also, to make debugging easier in futurer, PaStA data gets ordered after adding
new PaStA data which results from the filtering of patchstack mails.

Signed-off-by: Jakob Kronawitter <kronawij@fim.uni-passau.de>