Ensure sorting of commit-count and LOC-count data.frames
Due to weird circumstances, tests are failing only on R 3.3, indicating
string mismatches in all commit-count and LOC-count data.frames [1]. The
reason for these failures is that the row ordering of the respective
data.frames is not deterministic for authors exhibiting the same
frequency value.

To fix this problem, the resulting data.frames of all corresponding
functions are ordered by the author and committer names, respectively.

Props to @hechtlC for a valuable discussion.

[1] https://travis-ci.com/se-passau/coronet/jobs/278309743

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>