Change process of calculating EDCPTD centrality
The process of how EDCPTD centrality is computed is changed. In the
first step the author network for each relation is build. In the second
step the author networks are converted to a forth-order tensor and in
the third step EDCPTD centrality is computed.

Changes are based on the reviews from @clhunsen, @bockthom and
@ecklbarb.

Signed-off-by: Signed-off-by: Anselm Fehnker <anselm@muenster.de>