Fix function 'split.networks.time.based' regarding sliding windows
When calling the function 'split.networks.time.based' (which splits several
networks at the same time), the 'sliding.window' parameter was useless:
This function, in turn, calls 'split.network.time.based' for each network,
specifying fixed bins and passing the 'sliding.window' parameter to it.
However, if bins are given to this function, the 'sliding.window' parameter
is ignored. To fix this, create the additional steps of the sliding window
approach here manually in a similar way as it is done in
'split.network.time.based'. With this fix, it is possible to use sliding
windows when splitting multiple networks simultaneously.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>