Adjust package installation script to fix travis builds
Travis builds currently fail (for a few weeks) due to two different reasons:

On the one hand, the log output exceeds the maximum log length of travis,
resulting in failing builds. This is prevented by setting the parameters
`verbose` and `quiet` to `TRUE`, as then the package compilation log is
not printed to the log but the result of package installation is clearly
perceptible in the log.

On the other hand, travis installs lots of packages which are not necessary
as they are just suggestions of other packages but not necessary dependencies.
Installing such unnecessary packages my lead to problems. This is
circumvented by setting the `dependencies` parameter to `NA` as this results
in installing only necessary dependencies and imports, whereas the former
`TRUE` value of this parameter also led to installing suggested but
unnecessary packages.

Those two fixes hopefully makes travis builds successful again.

Props to @clhunsen for experimenting with the parameters and proposing
the first part of this commit regarding the `verbose` and `quiet`
parameters of the `install.packages` function.

This commit addresses parts of #161.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>