Fix computation of overlapping ranges
When computing overlapping ranges, sometimes we ended up in having one more
range in the end than needed. This led to the having ranges covering
almost the same time span. Consider the following example:

When constructing overlapping ranges of three month ranges which overlap
fifty-fifty, we might end up in the following scenario:

2020-06-27 ——— 2020-09-26 — 2020-11-11 (2020-12-25)
       2020-08-12 ————————— 2020-11-11

There is one range from 2020-06-27 to 2020-09-26,
         one range from 2020-08-12 to 2020-11-11,
     and one range from 2020-09-26 to 2020-12-25.

However, when assuming that the last action in the data takes place on
2020-11-11, then we would have two ranges covering the last range: one
complete ranges and half a range:

2020-08-12 to 2020-11-11 and 2020-09-26 to 2020-11-11.

As can be clearly seen (and also above in the visual example), the range
2020-09-26 to 2020-11-11 is superfluous as it does not even represent a full
range and it fully overlapps with the previous range. Hence we don't
need this half range and can just end up with the following ranges:

2020-06-27 ——— 2020-09-26
       2020-08-12 ————————— 2020-11-11

To achieve this, determine the number of complete ranges by rounding
downwards instead of rounding to the nearest integer.
However, if the complete time span does not even include one complete
range, set the number of complete ranges to 1 because multiplying with 0
further down in the function will end up having no range at all (even
not an incomplete one). However, the implementation of this corner case
works correctly as there already exist some extra checks for this
special case.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>