Fix off-by-one error in data cutting
The function `get.data.cut.to.same.date` cuts the data of multiple data
sources to a common date range for all the specified data sources.
However, in the previous implementation, there was an off-by-one error,
caused by excluding end dates in data splitting:

If a range is specified, the end date of the range is exclusive (while
the start of the range is inclusive). As the data cutting functionality
uses the split functionality, the specified end date was excluded
whereas it should be included, as the following example shows:

Assume there are two data sources d1 and d2. Let d1 contain the time
stamps t1 and t4, and let d2 contain the time stamps t2 and t3, with
t1 < t2 < t3 < t4. Then a visualization of this example on a time line
could be as follows:

d1:  |                                          |
d2:      |                                   |
    t1  t2                                  t3 t4

According to the proceedure of data cutting, we look for a common data
range. It can be easily seen that the interval [t2,t3] completely covers
data from both data sources d1 and d2. Hence, all the data needs to be
cut to this interval [t2,t3].
However, what the former implementation did was the following: t2 and t3
are determined as start and end of the range the data has to be cut to.
As the cutting functionality uses the splitting functionality, t2 is
used as start of the range and t3 as used as the end of the range. In
here the error comes in: As the end of ranges is exclusive, t3 is not
included in the cut data, resulting in interval [t2, t3).
To include t3 into the cut data, we need to use t3+1 as the end of the
range to cut, as this leads to [t2,t3] then.

In this commit, this off-by-one error at the end of the data cutting
range is fixed by adding 1 second to the end.
In addition, all the corresponding tests in the test suite (i.e., data
cutting and network cutting) are adjusted to incorporate the end of the
cutting range.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>