Fix function 'split.networks.time.based' regarding sliding windows
When calling the function `split.networks.time.based` (which splits several
networks at the same time), the `sliding.window` parameter was useless:
This function, in turn, calls `split.network.time.based` for each network,
specifying fixed bins and passing the `sliding.window` parameter to it.
However, if bins are given to this function, the `sliding.window` parameter is
ignored. To fix this, create the overlapping ranges need to get sliding windows
and call `split.network.time.based.by.ranges` to get the networks split using
sliding windows.

With this fix, it is possible to use sliding windows when splitting multiple
networks simultaneously.

Props to @clhunsen for suggesting to use `construct.overlapping.ranges` and
`split.network.time.based.by.ranges` for this fix.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>
Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>