Add tests for sliding windows
Add a completely new test module 'test-split-sliding-window.R', which is
based on the exisiting test module 'test-split.R'. The new test module
contains almost the same test setups as the exisiting test module, but
it tests all the splitting functions only using sliding windows.
In addition, some new test scenarios are added to test certain corner
cases with respect to the last range of sliding-window splitting.

In particular, the following functions are tested with sliding windows:

- split.data.time.based
- split.data.activity.based
- split.network.time.based
- split.network.activity.based

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>