Add more sliding-window tests & add library for parameterized tests
In this commit, additional tests for sliding-window splitting are added.
For example, tests for the function 'split.networks.time.based'.

As these tests don't need their expected outcomes to be determined
manually, but rely on the outcomes of other, already tested functions,
it is not necessary to completely copy existing tests and have them
duplicated for different values of `sliding.window`. Hence, this is a
good example for having parameterized tests:

Parameterized tests should succeed independently of which value certain
parameters take. Hence, we can just implement the test once and run it
twice, for example, the first run with `sliding.window = TRUE` and the
second run with `spliding.window = FALSE`.

However, es our standard test library 'testthat' does not provide the
possiblity to run parameterized tests, we need to add an additiona
library to be able to run parameterized tests: the library 'patrick'.

For more details on 'patrick', see: https://github.com/google/patrick

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>