Fix missing layout when plotting networks
For unknown reasons, setting the layout for ggraph, which is used for
plotting networks, has changed. To avoid errors due to a missing or
wrongly set layout, the corresponding implementation in function
'plot.get.plot.for.network' is adjusted in to ways:

1) When creating a ggraph object, we already need to explicitly specify
   the layout we want to use. As we rely on igraph and its layouts,
   we use 'layout = "igraph"' and then we have to specify the concrete
   layout algorithm from igraph (as 'algorithm' paramater).

2) Opposed to previous layout specification, now a layout name has to be
   stated instead of naming a layout function. Unfortunately, there is
   no official list of supported igraph layouts. Hence, we need to rely
   on the following table (which comes from python and not from R):
   https://igraph.org/python/doc/tutorial/tutorial.html#layout-algorithms
   Notice that the layout algorithm needs to be given as string in form of
   the 'short name' given in this table. However, if there are multiple
   short names specified in the table, only one of them might actually
   work for ggraph.

Now as before, the default layout is 'kk' (kamada kawai). If one wants
to use another igraph layout, just set the graph attribute 'layout' to
the short name of the layout and then this one will be used for creating
the ggraph object.

This commit fixes issue #186.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>