Fix various bugs regarding the PaStA data
First, fix the copy-paste error where the existence of mail data was checked
when adding the PaStA data to the mail data. Second, fix the duplication of
revision.set.ids that happens when merging the PaStA data to the commit or
mail data. Last, throw out PaStA items that contain message ids or commit
hashes that do not appear in the mail or commit data.

Signed-off-by: Christian Hechtl <hechtl@cs.uni-saarland.de>