Fix filtering patchstack mails if there are no mails
If there are no mails (e.g., after splitting the data to a certain
range), the function `filter.patchstack.mails()` prints warnings and
sets the mail data to NULL, whereas the mail data should stay as created
by `create.empty.mails.list()`. When then accessing certain columns of
the mail data, this leads to an error. To prevent this, don't perform
the patchstack mail filtering if there are no mails - just return the
mail data as they have been before.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>
Signed-off-by: Christian Hechtl <hechtl@cs.uni-saarland.de>