Replace build status badges: Use drone instead of Travis CI
As we have enabled drone.io for CI, also use the corresponding badges to
show the build status on our CONTRIBUTING.md.

In addition, remove the word "Travis" from the README.md as we won't use
Travis CI anymore and it is not necessary to state at the README.md
which CI tool we use.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>