Fix merging by hash instead of commit.id
As commit.id was the first column of the data frame anyway,
merging has not changed the order. But when using the hash column
it is taken as the first colum of the resulting data frame.
Change the order of the columns in order to not break anything
that relies on the order.

See #180

Signed-off-by: Niklas Schneider <s8nlschn@stud.uni-saarland.de>