Adjust README and add config file for GH-pages website
Use our README.md file for automatically generating a GH-pages website
for coronet.

To achieve that, a _config.yml is necessary in which we specify how the
website should be generated.

Also links in the README.md should be relative links, otherwise the
links will only work in the README rendered on the GitHub repo, but not
on the generated GH-pages website.

In addition, also fix some outdated links, add missing section to table
of contents, etc.

Add favicon to README.md, CONTRIBUTING.md, and NEWS.md.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>