Fix problems in CI pipeline for R-3.3
Our CI pipeline does currently not work for "R-3.3" (but for all the other
R versions in the pipeline). The reason for that is the following:
As of 2021-01-26, there is a new version of package `memoise` (2.0.0),
which is imported by `RSQLite`, which in turn is imported by `sqldf`.
The new version of `memoise` imports package `cachem`, which in turn
imports package `fastmap`. Using all these packages is not a problem per
se, they are compatible with R 3.3.3. However, the docker container
`r-base:3.3.3`, which we use in our pipeline "R-3.3", uses g++ version
6.3.0-9, which contains a bug. Due to the bug in the compiler, `fastmap`
cannot be compiled and installed in tje `r-base:3.3.3` docker container,
letting the CI pipeline fail.

To circumvent this problem but still keep "R-3.3" in our CI pipeline, we
now use docker image `r-base:3.3.2` instead, as there is another g++
version installed in this docker image. With that, the CI pipeline should
succeed for "R-3.3" again.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>