Add checks for artifact networks and vectorize 'get.artifacts'
As the default restrict parameter for 'metrics.centrality' depends
on the network's vertices, add checks and correpsonding calls to
'get.artifact' if a network contains artifact vertices.
Vectorize 'get.artifact' in 'util-data.R' in order to be able to
process and return artifacts from different data sources. The function
now takes a vector of data sources (or still a single one) and
returns all correpsonding artifacts.

See #195

Signed-off-by: Niklas Schneider <s8nlschn@stud.uni-saarland.de>