Add manual default check to 'get.artifacts'
Now that there can be multiple values in the argument, the
'match.arg.or.default' function shows a different behaviour. Instead of
returning only the first element of the default vector when no argument
is passed, it returns the whole vector.
Add manual checks for a missing argument and set it to the first element
in this case. Inspired from the implementation of 'match.arg.or.default'

See #195

Signed-off-by: Niklas Schneider <s8nlschn@stud.uni-saarland.de>