Add range data flag to constructor and adapt calls in util-bulk.range
Add the 'built.from.range.data' to the constructor of the 'RangeData'
class which sets its private field.
Add a getter method 'is.built.from.range.data' that returns the value of
the private attribute.
Adapt the calls to the constructor in 'util-bulk.R' since these are the
only cases where the flag needs to be set true.

Signed-off-by: Niklas Schneider <s8nlschn@stud.uni-saarland.de>