Add private vectors to ProjectData containing names of read data sources
Add private fields 'read.data.sources' and
'read.additional.data.sources' along with a getter for each of them to
the ProjectData class.
The getters for the (additional) data sources will now check whether
their respective source is already in the appropriate vector. If not, it
will trigger the reading process and the name of the source is added to
the vector.

This is needed because there is no way to check which data sources have
been read at a given point in time without triggering the reading
process.

See #177

Signed-off-by: Niklas Schneider <s8nlschn@stud.uni-saarland.de>