Remove warning assignment from test suite initialization
As of R version 4.1, the base environment is locked and bindings to the
base environment cannot be added or removed any more. Hence, assigning
'last.warning' to NULL results in an error, as 'last.warning' belongs to
the base environment.

To make the test suite run also on R version 4.1, the assignment that clears
'last.warning' ist removed. This should not affect the tests in our CI
pipeline as ususually they are run in a new environment.
For now, there is no possibility any more to clear previous warnings (if
existent).

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>