Add 'tzone' attribute to empty POSIXct vectors
With the release of R version 4.1.0, the R developers fixed a bug with
the 'tzone' attribute of empty POSIXct vectors. In previous R versions,
when calling `as.POSIXct()` on an empty vector `c()` or on an `NA`
value, no `tzone` attribute was stored, which was inconsistent as when
calling `as.POSIXct()` on an actual value the `tzone` attribute was
always added. Since R version 4.1.0, the `tzone` attribute is also added
on empty vectors and `NA` values. (See also the official R news, Section
"Changes in R 4.1.0", Subsection "Bug fixes":
https://cran.r-project.org/doc/manuals/r-release/NEWS.html )

However, this bugfix in R version 4.1.0 breakes our test suite, as in
some situations empty date vectors are created or `NA` values are passed
as default value for a date, but no time zone is set in these cases. As
this leads to the failure of certain tests, the time zone needs to be
correctly set to UTC whenever `NA` dates or empty date vectors are
created. This is fixed by either setting the time zone attribute via
`lubridate::with_tz` or by calling `get.date.from.string` on `NA`
values.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>