Add new parameter to get.cached.data.sources and adapt checks for splits
When having been created by reading (Codeface) ranges, a RangeData
object must not split data after reading as the range it is given has
no dates associated with it, but only the name of the read range.

Split the filtered data sources into a separate vector in
'get.cached.data.sources' and add a toggle parameter for these to make
it more consistent.

Signed-off-by: Niklas Schneider <s8nlschn@stud.uni-saarland.de>