Add locking functionality for main data sources
Add locks in form of a config parameter for each of commits, mails, and
issues in the ProjectConf class.
The getters of these main data sources will now check whether the lock
is set before triggering a re-read.
Exclude those new parameters from resetting the envíronment; if they
would, it would be nonesense to have them.
Change tests in test-networks-covariates to use the locks on issues
because the tests expect no issues and before, the getters would read
them anyway which lead to failures.

Signed-off-by: Niklas Schneider <s8nlschn@stud.uni-saarland.de>