Fix logging statement when printing a Conf object
When printing a Conf object, the info "Network configuration:" was printed in
the first line. However, this is actually wrong as this print function is used
from both the `NetworkConf` and the `ProjectConf` object, as well as the `Conf`
class itself. To avoid printing wrong information and also to avoid code
duplication, just print "Configuration:" from now on, as the `Conf` class does
not know which actual class the object at which this function is called belongs
to.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>