Fix generation of Codeface range directory names
Directory names for Codeface ranges can be generated via revision names or
commit hashes. If commit hashes are used (which are 40 characters long),
Codeface only uses the first 6 characters in the directory names. This was
already implemented correctly, but the check has used the wrong function: To
check if the revision name is 40 characters long, use the `nchar` function
instead of the `length` function (as the latter always returns 1 in this case).
This is a follow-up commit for b3e29472a57e26935a31645b96fbef7d7785c25a, which
has already been introduced with version 3.5 of coronet.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>