Fix plotting an empty network
If an empty network is passed to the plot function `plot.network`, vertex
attributes `kind` and `vertex.type` have not been present in the created empty
network. As both attributes are used during the plotting function, this lets
plotting an empty network fail. This commit adjusts the plotting function such
that we can plot empty networks without any errors.

In addition, fix a wrong comment in function `add.attributes.to.network`.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>