Fix bug in issue-related `add.vertex.attribute.issue.*` functions
When not passing a specific value for parameter `issue.type` to all the
`add.vertex.attribute.issue.*` functions, the vector `c("all", "pull.requests",
"issues")` is automatically used as default. However, as only one of the
vector's elements is allowed, this leads to problems when passing this parameter
to other functions in turn. To fix this, add some missing `match.arg` calls that
make sure that only one value (the actual default "all") is used for the
`issue.type` parameter.

This is a follow-up commit for PR #194 and commit
eb4f649c04155e22195627072e0f08bb8fe65dc4.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>