Fix erroneous function calls in the ProjectData class
Within functions `set.project.conf.entry` and `update.project.conf`, the
synchronicity data is updated via calling the setter and the getter for the
synchronicity data. As both getter and setter are public functions onf the
ProjectData class, calling `private$...` does not work. Instead, we have to use
`self$...`, which holds for public functions.

This is a follow-up commit for PR #201 and commit
73009d9c2a0c29b6c760534f66c4b19bb5b8cb45.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>