Reduce amount of multiple calls to 'update.synchronicity.data()'
When resetting a ProjectConf entry that is related to the synchronicity data,
the corresponding update function is called. However, sometimes, the update
function is called multiple times when synchronicity data is already cached:
- The setter calls the update function (if synchronicity is configured).
- The getter calls the update function (if synchronicity is configured).
- The update function is called manually.
So, the manual function call is not necessary if the getter was already called
before (and both synchronicity data is configured and cached).

This is another follow-up commit for PR #201 and commit
73009d9c2a0c29b6c760534f66c4b19bb5b8cb45.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>